+++ packages/@formik/reducer-refs/src/hooks/useField.tsx
@@ -1,20 +1,9 @@
 import * as React from 'react';
-import {
-  isFunction,
-  isEmptyChildren,
-  isObject,
-  FormikProps,
-  GenericFieldHTMLAttributes,
-  FieldMetaProps,
-  FieldHelperProps,
-  FieldInputProps,
-  FieldValidator,
-  selectGetFieldMeta
-} from '@formik/core';
-import invariant from 'tiny-warning';
-import { useFormikApi } from './hooks/useFormikApi';
-import { FormEffect } from './types';
+import { FormEffect } from '../types';
 import isEqual from 'react-fast-compare';
+import { FieldHelperProps, FieldInputProps, FieldMetaProps, FieldValidator, FormikProps, GenericFieldHTMLAttributes, isObject, selectGetFieldMeta } from '@formik/core';
+import { useFormikApi } from './useFormikApi';
+import invariant from 'tiny-warning';
 
 export interface FieldProps<V = any, FormValues = any> {
   field: FieldInputProps<V>;
@@ -105,7 +94,8 @@ export function useField<Val = any>(
   const [fieldMeta, setFieldMeta] = React.useState(fieldMetaRef.current);
 
   const maybeUpdateFieldMeta = React.useCallback<FormEffect<any>>((formikState) => {
-    // we could use formikApi.getFieldMeta... but is that cheating?
+    // we could use formikApi.getFieldMeta... but is that correct?
+    // I think we should use the value passed to this callback
     const fieldMeta = selectGetFieldMeta(() => formikState)(fieldName);
 
     if (!isEqual(fieldMeta, fieldMetaRef.current)) {
@@ -149,85 +139,4 @@ export function useField<Val = any>(
     fieldMeta,
     getFieldHelpers(fieldName),
   ];
-}
-
-export function Field(rawProps: FieldAttributes<any>) {
-  const {
-    render,
-    children,
-    as: is, // `as` is reserved in typescript lol
-    component,
-    ...props
-  } = rawProps;
-
-  if (__DEV__) {
-    // eslint-disable-next-line react-hooks/rules-of-hooks
-    React.useEffect(() => {
-      invariant(
-        !render,
-        `<Field render> has been deprecated and will be removed in future versions of Formik. Please use a child callback function instead. To get rid of this warning, replace <Field name="${name}" render={({field, form}) => ...} /> with <Field name="${name}">{({field, form, meta}) => ...}</Field>`
-      );
-
-      invariant(
-        !(is && children && isFunction(children)),
-        'You should not use <Field as> and <Field children> as a function in the same <Field> component; <Field as> will be ignored.'
-      );
-
-      invariant(
-        !(component && children && isFunction(children)),
-        'You should not use <Field component> and <Field children> as a function in the same <Field> component; <Field component> will be ignored.'
-      );
-
-      invariant(
-        !(render && children && !isEmptyChildren(children)),
-        'You should not use <Field render> and <Field children> in the same <Field> component; <Field children> will be ignored'
-      );
-      // eslint-disable-next-line
-    }, []);
-  }
-
-  const [field, meta] = useField(props);
-  const formik = useFormikApi();
-
-  const legacyBag = { field, form: formik };
-
-  if (render) {
-    return render({ ...legacyBag, meta });
-  }
-
-  if (isFunction(children)) {
-    return children({ ...legacyBag, meta });
-  }
-
-  if (component) {
-    // This behavior is backwards compat with earlier Formik 0.9 to 1.x
-    if (typeof component === 'string') {
-      const { innerRef, ...rest } = props;
-      return React.createElement(
-        component,
-        { ref: innerRef, ...field, ...rest },
-        children
-      );
-    }
-    // We don't pass `meta` for backwards compat
-    return React.createElement(
-      component,
-      { field, form: formik, ...props },
-      children
-    );
-  }
-
-  // default to input here so we can check for both `as` and `children` above
-  const asElement = is || 'input';
-
-  if (typeof asElement === 'string') {
-    const { innerRef, ...rest } = props;
-    return React.createElement(
-      asElement,
-      { ref: innerRef, ...field, ...rest },
-      children
-    );
-  }
-
-  return React.createElement(asElement, { ...field, ...props }, children);
 }
