+++ packages/@formik/reducer-refs/src/index.tsx
@@ -1,11 +1,15 @@
+export * from './components/ErrorMessage';
+export * from './components/FastField';
+export * from './components/Field';
+export * from './components/FieldArray';
+export * from './components/Form';
+export * from './components/Formik';
+export * from './connect';
+export * from './contexts/FormikApiContext';
+export * from './contexts/FormikContext';
+export * from './hooks/useField';
 export * from './hooks/useFormik';
 export * from './hooks/useFormikApi';
-export * from './contexts/FormikContext';
-export * from './Formik';
-export * from './Field';
-export * from './Form';
+export * from './hooks/useFormikContext';
+export * from './hooks/useFormikState';
 export * from './withFormik';
-export * from './FieldArray';
-export * from './connect';
-export * from './ErrorMessage';
-export * from './FastField';
